requirements.txt
flask==3.0.0
python-dotenv==1.0.0
flask-login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
email_validator==2.1.0.post1
flask-ipban==1.1.5
requests==2.31.0
